+++ bluer_sbc/README/aliases.py
@@ -1,0 +1,10 @@
+docs = [
+    {
+        "path": f"./docs/aliases/{item}.md",
+    }
+    for item in [
+        "camera",
+        "hardware",
+        "rpi",
+    ]
+]
