+++ bluer_sbc/__init__.py
@@ -4,7 +4,7 @@ ICON = "🌀"
 
 DESCRIPTION = f"{ICON} AI for single board computers and other designs."
 
-VERSION = "8.214.1"
+VERSION = "8.218.1"
 
 REPO_NAME = "bluer-sbc"
