+++ bluer_sbc/designs/shelter.py
@@ -1,8 +1,15 @@
 from bluer_objects import README
+from bluer_objects.README.items import ImageItems
+from bluer_objects.README.consts import assets_path
 
 from bluer_sbc.designs.consts import assets2
 
 image_template = assets2 + "shelter/{}?raw=true"
+
+assets2_shelter = assets_path(
+    suffix="shelter",
+    volume=2,
+)
 
 marquee = README.Items(
     [
@@ -21,5 +28,12 @@ items = README.Items(
             "name": "",
         }
         for index in range(4)
-    ],
+    ]
+    + [
+        ImageItems(
+            {
+                f"{assets2_shelter}/20251005_180841.jpg": "",
+            }
+        )
+    ]
 )
